requirements.txt
qrcode
ecdsa
colorama
pymerkle
pycryptodomex